refactor: extract date store factory from GlobalDateStoreProvider
Move the global date store implementation into a reusable makeDateStore
factory under src/providers/DateStore/shared, parameterized by initial
range. GlobalDateStoreProvider becomes a thin instantiation that
re-exports the exact same names from its new home at
src/providers/DateStore/GlobalDateStore. Import paths updated across
internal consumers; no public API or behavior changes.
