+++ src/components/DateSelection/CombinedDateSelection.tsx
@@ -1,4 +1,4 @@
-import type { DateSelectionMode } from '@providers/GlobalDateStore/GlobalDateStoreProvider'
+import type { DateSelectionMode } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
 import { GlobalDateSelection } from '@components/DateSelection/GlobalDateSelection'
 import { GlobalMonthPicker } from '@components/GlobalMonthPicker/GlobalMonthPicker'
