+++ src/components/Trips/TripsView/ResponsiveTripsView.tsx
@@ -7,7 +7,7 @@ import type { Trip } from '@schemas/trip'
 import { BREAKPOINTS } from '@utils/screenSizeBreakpoints'
 import { useListTrips } from '@hooks/api/businesses/[business-id]/mileage/trips/useListTrips'
 import { useAutoResetPageIndex } from '@hooks/utils/pagination/useAutoResetPageIndex'
-import { useGlobalDateRange } from '@providers/GlobalDateStore/GlobalDateStoreProvider'
+import { useGlobalDateRange } from '@providers/DateStore/GlobalDateStore/GlobalDateStoreProvider'
 import { useCurrentTripsPage, useTripsTableFilters } from '@providers/TripsRouteStore/TripsRouteStoreProvider'
 import { type DefaultVariant, ResponsiveComponent } from '@ui/ResponsiveComponent/ResponsiveComponent'
 import { DataState, DataStateStatus } from '@components/DataState/DataState'
